context deprecation
This lets you send lwsws a SIGHUP and have it reload the configuration.

It does this without affecting existing connections, which continue to
run under the old configuration until no more old connections are left.

New connections are made to the new context using the new configuration.

Notes:

1) Only lwsws has the code to handle the signal, you can cut and paste it
into your own server if needed

2) lws_context_destroy() has been split into two pieces... the reason for
the split is the first part closes the per-vhost protocols, but since
they may have created libuv objects in the per-vhost protocol storage,
these cannot be freed until after the loop has been run.

That's the purpose of the second part of the context destruction,
lws_context_destroy2().

For compatibility, if you are not using libuv, the first part calls the
second part.  However if you are using libuv, you must now call the
second part from your own main.c after the first part.

3) There are also significant updates to the system-status plugin, so you
can clearly see what is going on with the deprecated contexts.